fix(compose): no var interpolation + path resolution
Due to consolidating the InfluxDBv1/v2 related env vars into
`.env` the playbook generates a Compose file with no values set
for any environment variables that need to be interpolated via
the `.env` file. In order to mitigate this, the compose validation
+ generation tasks do not interpolate the env vars and remove
path resolution too. With this fix, the files generated can then
rely on `.env` file even if the file generated after validation.

fixes #102

Signed-off-by: Shantanoo 'Shan' Desai <shantanoo.desai@gmail.com>